+++ src/Kubia/Camunda/CamundaInitiator.php
@@ -1,11 +1,12 @@
 <?php
 
 namespace Kubia\Camunda;
-
+use Camunda\Entity\Request\ProcessInstanceRequest;
+use Camunda\Service\ProcessInstanceService;
 use Camunda\Entity\Request\ProcessDefinitionRequest;
 use Camunda\Service\ProcessDefinitionService;
 use PhpAmqpLib\Message\AMQPMessage;
-use Quancy\Logger\Logger;
+use Kubia\Logger\Logger;
 
 /**
  * Class CamundaInitiator
@@ -16,24 +17,40 @@ class CamundaInitiator extends CamundaBaseConnector
     /** @var string */
     public $logOwner = 'bpm-initiator';
 
+    /** @var bool */
+    public $startAllowed = true;
+
     /**
-     * Callback
-     * @param AMQPMessage $msg
+     * Check running process instances with current business key
      */
-    public function callback(AMQPMessage $msg): void
+    public function isProcessInstanceAlreadyStarted(): void
     {
-        Logger::log(sprintf("Received %s", $msg->body), 'input', $this->rmqConfig['queue'], $this->logOwner, 0 );
+        // count running process instances
+        $processInstanceRequest = (new ProcessInstanceRequest())
+            ->set('businessKey', $this->headers['camundaBusinessKey']);
 
-        // Set manual acknowledge for received message
-        $this->channel->basic_ack($msg->delivery_info['delivery_tag']); // manual confirm delivery message
+        $processInstanceService = (new ProcessInstanceService($this->camundaUrl))
+            ->getListCount($processInstanceRequest);
 
-        // Update variables
-        $this->message = json_decode($msg->body, true);
-        $this->headers = $this->message['headers'] ?? null;
+        // if process already exist
+        if(
+            (int)$processInstanceService->count > 0 &&
+            isset($this->headers['camundaProcessUnique']) &&
+            filter_var($this->headers['camundaProcessUnique'], FILTER_VALIDATE_BOOLEAN)
+        ) {
+            // disallow start
+            $this->startAllowed = false;
 
-        // Validate message
-        $this->validateMessage();
+            if($this->msg->has('correlation_id') && $this->msg->has('reply_to')) {
+                $this->sendSynchronousResponse($this->msg);
+            }
+        }
+    }
 
+    /**
+     * Start process instance
+     */
+    public function startProcessInstance(): void {
         // Update variables
         $this->updatedVariables['message'] = [
             'value' => json_encode($this->message),
@@ -41,7 +58,8 @@ class CamundaInitiator extends CamundaBaseConnector
         ];
 
         $processDefinitionRequest = (new ProcessDefinitionRequest())
-            ->set('variables', $this->updatedVariables);
+            ->set('variables', $this->updatedVariables)
+            ->set('businessKey', $this->headers['camundaBusinessKey']);
 
         $processDefinitionService = new ProcessDefinitionService($this->camundaUrl);
         $key = $this->camundaConfig['prefix'] . $this->headers['camundaProcessKey'];
@@ -64,6 +82,34 @@ class CamundaInitiator extends CamundaBaseConnector
                 $processDefinitionService->getResponseContents()->message ?? $this->requestErrorMessage
             );
             Logger::log($logMessage, 'input', $this->rmqConfig['queue'], $this->logOwner, 1 );
+        }
+    }
+
+    /**
+     * Callback
+     * @param AMQPMessage $msg
+     */
+    public function callback(AMQPMessage $msg): void
+    {
+        Logger::log(sprintf("Received %s", $msg->body), 'input', $this->rmqConfig['queue'], $this->logOwner, 0);
+
+        // Set manual acknowledge for received message
+        $this->channel->basic_ack($msg->delivery_info['delivery_tag']); // manual confirm delivery message
+
+        $this->msg = $msg;
+
+        // Update variables
+        $this->message = json_decode($msg->body, true);
+        $this->headers = $this->message['headers'] ?? null;
+
+        // Validate message
+        $this->validateMessage();
+
+        // Check running process instances with current business key
+        $this->isProcessInstanceAlreadyStarted();
+
+        if($this->startAllowed) {
+            $this->startProcessInstance();
         }
     }
 }