+++ CHANGELOG.md
@@ -5,7 +5,16 @@ BPM Initiator on PHP. Using to start business process in Camunda BPM.
 ## [Unreleased]
 
 ### Planned
+
 - Nothing
+
+## [0.2] - 2020-01-17
+
+### Added
+
+- Inheritance from CamundaBaseConnector
+- Check running process instances with current business key
+- Examples for test send synchronous and asynchronous request for initialize process
 
 ## [0.1] - 2019-12-20
 
@@ -18,5 +27,6 @@ BPM Initiator on PHP. Using to start business process in Camunda BPM.
 - All ticks timeout moved to config
 - Format transit messages from Rabbit MQ changed from `String` to `Json`
 
+[unreleased]: https://gitlab.com/quancy-core/bpm-initiator/-/tags/0.2
 [unreleased]: https://gitlab.com/quancy-core/bpm-initiator/-/tags/0.1
 [0.1]: https://gitlab.com/quancy-core/bpm-initiator/-/tags/0.1